# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking workouts, exercises, and fitness routines",
        "title": "GetFit API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.yoursite.com/support",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/search/{exerciseID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find an exercise using its ID without user context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Search for an exercise by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise information",
                        "schema": {
                            "$ref": "#/definitions/store.Exercise"
                        }
                    },
                    "400": {
                        "description": "Invalid exercise ID format",
                        "schema": {}
                    },
                    "404": {
                        "description": "Exercise not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to fetch exercise",
                        "schema": {}
                    }
                }
            }
        },
...
        "/users/{userID}/exercise": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all exercises created by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Get all exercises for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exercises",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Exercise"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to fetch exercises",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a custom exercise for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Create a new exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Exercise information",
                        "name": "exercise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.Exercise"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Exercise created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing required fields",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to create exercise",
                        "schema": {}
                    }
                }
            }
        },
        "/users/{userID}/exercise/{exerciseID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a specific exercise by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Get exercise by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise information",
                        "schema": {
                            "$ref": "#/definitions/store.Exercise"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {}
                    },
                    "404": {
                        "description": "Exercise not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to fetch exercise",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove an exercise from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Delete an exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to delete exercise",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing exercise's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Update an exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exerciseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated exercise information",
                        "name": "exercise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.updateExercisePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing fields",
                        "schema": {}
                    },
                    "404": {
                        "description": "Exercise not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to update exercise",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.updateExercisePayload": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "equipment": {
                    "type": "string"
                },
                "expected_version": {
                    "type": "integer"
                },
                "force": {
                    "type": "string"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "level": {
                    "type": "string"
                },
                "mechanic": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primaryMuscles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "secondaryMuscles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "store.Exercise": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "equipment": {
                    "type": "string"
                },
                "force": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_custom": {
                    "type": "boolean"
                },
                "level": {
                    "type": "string"
                },
                "mechanic": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primaryMuscles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "secondaryMuscles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}